[Type checker] Move override checking into its own file.
There is a lot of code involved in override checking; move it into its own
source file. NFC
